fix: budget None-guard, escalation uuid reuse, shared control-char regex
Addresses external-reviewer findings on PR #2281:
- budget/rebalance._budget_percent: treat explicit None as 0.0, same as
  a missing key (Gemini); correct the now-stale type-only-cast docstring.
- security/service_escalation: keep the generated UUID object and reuse
  it for ApprovalItem.id instead of re-parsing the string form (Gemini).
- Extract the shared control/formatting CONTROL_CHAR_RE into
  security/_shared_patterns and consume it from safety_classifier,
  information_stripper, and llm_evaluator so the codepoint set lives in
  one place (CodeRabbit).
- Regenerate feature_index / codebase_map / architecture_report.
