+++ data/codebase_map.json
@@ -2522,7 +2522,7 @@
     },
     {
       "kind": "code",
-      "loc": 118,
+      "loc": 121,
       "loc_cap": 500,
       "module": "src/synthorg/budget/rebalance.py",
       "owning_feature": "budget"
@@ -17047,6 +17047,13 @@
     },
     {
       "kind": "code",
+      "loc": 33,
+      "loc_cap": 500,
+      "module": "src/synthorg/security/_shared_patterns.py",
+      "owning_feature": "security"
+    },
+    {
+      "kind": "code",
       "loc": 32,
       "loc_cap": 500,
       "module": "src/synthorg/security/action_type_mapping.py",
@@ -17201,7 +17208,7 @@
     },
     {
       "kind": "code",
-      "loc": 86,
+      "loc": 64,
       "loc_cap": 500,
       "module": "src/synthorg/security/information_stripper.py",
       "owning_feature": "security"
@@ -17544,7 +17551,7 @@
     },
     {
       "kind": "code",
-      "loc": 497,
+      "loc": 486,
       "loc_cap": 500,
       "module": "src/synthorg/security/safety_classifier.py",
       "owning_feature": "security"
@@ -17558,7 +17565,7 @@
     },
     {
       "kind": "code",
-      "loc": 137,
+      "loc": 138,
       "loc_cap": 500,
       "module": "src/synthorg/security/service_escalation.py",
       "owning_feature": "security"
