+++ data/feature_index.json
@@ -1199,6 +1199,6 @@
       ]
     }
   ],
-  "generated_at": "2026-06-08T20:35:41.325182Z",
+  "generated_at": "2026-06-08T21:21:26.463022Z",
   "schema_version": 1
 }
